roll our own crypto
Lets not make system calls to gpg for our key management. Instead we roll our
own crypto based on recommendation from the cryptopasta repo and sensible file
permissioning. This means that we cannot use GPG keys and GPG key management
tools, but it does mean that we can explicitly define our key system. And it is
a pretty simle system, so it shouldn't be too hard to implement in other
clients when necessary.

Also moved the key storage mechanism to pem to allow us to expand or modify it
in the future.

added constants in various places where it makes sense, probably more in future