BROKEN, WIP: refactoring state system
As I started writing the state machine in earnest I found that the current
implementation of the stat is far too clunky. It tries too hard to be both the
file system state and the internal state and the IPFS state all at the same
time. This is a lot to juggle, so I've started from the ground up rewriting the
state packge. Now there is logical state for the porgram to manipulate, and
those structs know how to save and load themselves to files, as well as publish
and retrieve themselves from IPFS. Should make life a bit easier going forward.
I still need to add tests to the State struct and tests for all of the
IPFS-publishing objects. That's still something to figure out: how to test IPFS
interactions. Will probably involve the IPTB.

Anyway, the daemon and init commands have not been updated to use the new state
package. They won't be updated until I have unit tests in place. So for now the
commands are almost certainly broken.